pyproject.toml
[tool.poetry]
name = "repl_python3_TNNGBOT"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
discord = "^2.3.2"
Flask = "^2.0.2"
discord-interactions = "^0.3.0"
python-dotenv = "^0.20.0"
pymongo = "^4.3.3"
typing-extensions = "^4.5.0"
requests = "^2.28.2"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
